make Filesystem a final class
